backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
sqlalchemy==2.0.31
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9
pandas==2.2.2
scikit-learn==1.5.0
pydantic-settings==2.3.4
alembic==1.13.2
python-dotenv==1.0.1